+++ DualSense4GTAV/Config/KeyConf.cs
@@ -1,0 +1,13 @@
+﻿using System.Windows.Forms;
+
+namespace DualSense4GTAV.Config
+{
+    public static class KeyConf
+    {
+        public static readonly Keys showMenu = Keys.F8;
+        public static readonly Keys showBatStat = Keys.F9;
+        public static readonly Keys toggleBatStat = Keys.F11;
+        public static readonly Keys showCommStat = Keys.F10;
+
+    }
+}
