--- DualSense4GTAV/Main/KeyConf.cs
@@ -1,13 +1,0 @@
-﻿using System.Windows.Forms;
-
-namespace DualSense4GTAV
-{
-  public static class KeyConf
-  {
-    public static readonly Keys showMenu     = Keys.F8;
-    public static readonly Keys showBatStat  = Keys.F9;
-    public static readonly Keys toggleBatStat  = Keys.F11;
-    public static readonly Keys showCommStat = Keys.F10;
-
-  }
-}
